Create solution.py
O(n) time solution with O(n) space for the result array, where n is the row index.

For i in the range of rows, create the next row with an extra slot. Go through the range from j to the end of the result array, and add the value to both the current index j and the next index (j+1) in the next row. Then, set the result to the next row when this is done.

Finally, return the result.