Create solution.py
O(n) time solution with O(1) space using one-pass through the array.

First, initialize the variables max1 and max2 (the largest and 2nd largest value placeholders) to 0, and min1 and min2 (the smallest and 2nd smallest value placeholders) to infinity.

For each number in the array, if n is greater than the 2nd max value, check if it is also greater than the 1st max value. If so, set max1 to the number and max2 to the previous max1 value. Otherwise, just set max2 to the number. Similarly, if n is less than the 2nd min value, check if it is also less than the 1st min value. If so, set min1 to the number and min2 to the previous min1 value. Otherwise, just set min2 to the number.

Now, return the 2 max values multiplied together minus the 2 min values multiplied together.