Create solution.py
O(n) time and O(1) space solution.

First, set the number of zeroes to 0, and the number of ones as the count of the '1' character in the string (first O(n) pass through the string).

Initialize the result to 0.

For i in the range up to the 2nd-to-last element, if the character is a 0, increment the number of zeroes. Otherwise, decrement the number of ones (since the left side has the '1' character now). Update the result as the max of itself and the sum of the number of zeroes and number of ones at the given iteration.

Finally, return the result.