Create solution.py
O(n) time and O(n) space solution using hashset.

First,  initialize a direction-to-coordinate dictionary for each of the 4 cardinal directions.

Initialize a visited set of visited coordinates and the origin (x, y) = (0, 0).

For each point p in path, first add the current (x, y) coordinate to the path. Calculate the deltas from the direction-coordinate mapping as dx and dy. Then, add this to the x and y to update the coordinates. If (x, y) after this update is already in the visited set, return true.

If the loop finishes without returning, return false.