Create solution.py
O(n) time solution with O(1) space.

First, initialize the count which counts whether the even indices are not 1 and odd indices are not 0: i.e., the alternating pattern of the form '101010....'. The deviations are in the count.

For each index in the length of the string, if the index is odd, increment the count if there is a '1' character in this position, otherwise 0. Similarly, if the index is even (else case), add 1 to the count if the character in this position is a '0', else 0.

Return the minimum of the count and the length minus the count. The length minus the count is the case for the other alternating pattern '01010101....' This algorithm accomplishes a one-pass way to calculate both patterns.