Create solution.py
O(n) time and O(26) = O(1) space solution for the character-to-index dictionary.

First, initialize the char to index dictionary and set the result initially to -1.

For each index/character pair in the string, if the character is not in the dictionary, add the character as the key with the index at which it occurs. This will be the first index at which this particular character occurs. Otherwise, set the result to be the maximum of itself and the current index - the index value of that character in the dictionary (which is its first occurrence) - 1 (in order to not include the end characters.

Finally, return the result.