Create solution_hm.py
O(n) time and O(n) space solution using a hashmap.

First, initialize the result array and the dictionary of the count of each numbers.

For each number in the range of 1 to n, if the count of that number in the dictionary is 0, add the value to the 2nd position of the result array. If the count is 2 (a duplicate), add it to the first position in the result array.

Return the result.