Create solution_inplace.py
O(n) time and O(1) space solution using in-place modifications of the input array.

First, initialize the result array with 0.

For each number in the array, get the absolute value of that number and assign to itself. If the number at the (n-1)-th index is negative, this is a duplicate number, so set the first value in the result array as this number (since this is a duplicate). Then, update the value at the (n-1)-th index with its negation.

Running again through the array after the previous negation update for each index/number pair, if the number is positive (which means it has been visited twice) and the index + 1 (i+1) is not the first value in the result array, this means this number is the missing number. Thus, update the result's 2nd value with i+1 and return the result. This will always return at some point.