Create solution_math.py
O(n) time and O(1) space solution using algebraic solution of sum and square of sums, and differencing it with the theoretical sum and square of sum of the numbers from 1 to n.

First, initialize the result array. and assign the length of the array. Let x and y equal 0, where x = d - m and y = d^2 - m^2, where d is the duplicate in the array and m is the missing value in the array. By using algebraic manipulation, the formula can be calculated for each number is a - b, where a is from the array and b is from the range (1 to n). Similarly, a^2 - b^2 is calculated. Since the order does not matter for adding these numbers, one pass may be done.

Then, the missing and duplicate are calculated by the formula where

d - m = x, d^2 - m^2 = y
d = m + x
(m + x)^2 - m^2 = m^2 + 2xm + x^2 - m^2 = x^2 + 2xm = y

thus, 

m = (y - x^2)  / (2x) and d = m + x

Then, return the results.