Create solution.py
O(1) time solution for each next() call and init() call, with O(n) space for the window size.

First, initialize the capacity (which is the window size), the current size (which counts how many elements have been added so far, with a maximum of the capacity), the current average (set to 0) and the window array that contains the window of the latest values of the size. Similarly, also set the initial position to 0 (the position in the array which contains the last value).

In the next() method, check if the capacity equals the current size. This means the window array is full. First, calculate the online average, subtracting out the oldest value in the window array, and adding the new value and dividing. Set the new value passed in at this current position in the window array. Increment the current position to the next index (which need to be modded in order to keep this circular array). Otherwise (when the current size is less than the capacity), calculate the online average by adding to the current sum the value, then dividing by the current size + 1. Next, increment the current size, and set the new value to this position. Increment the position, and mod it by the capacity. Finally, return the current average.