Create solution.py
O(n) time and O(1) in-place solution using two pointers.

First, get the length of the character array and set the left and right pointers to the end of the array.

While the pointers do not cross, swap the left and right pointer elements and move the pointers across.

Set index i to 0.

For index j in the range up to n, if the j is n or the character at j is a space character, set the left and right pointers to the i and the j - 1 indices. While the pointers don't cross, swap the characters at the left and right pointers, and increment/decrement the left/right pointers, respectively. After this while loop exits, set i to j + 1 in order to jump ahead to the next word.