Create solution.py
O(n) time and O(n) space from the list of string.

Convert the string to a list.

Set the left pointer to 0.

For the right pointer in the range up to the end of the list, if the right pointer element is a space or the end of the string, set a temp left and temp right variable as the left pointer and right pointer - 1. In the case the right pointer is at the end of the list, just set it to the right pointer. While the temp left and right pointers do not meet, swap the characters, and increment/decrement the temp left/right pointers. After the while loop, set the left pointer to right + 1 (in order to start after the space on the next word).

Finally, join the list of strings and return.