Create solution.py
O(n) time and O(n) space using a list of the string and a two pointer approach.

First, convert the string into a list of characters.

Create a vowel set of all the lower case vowels 'aeiou'.

Set the left and right pointers to the endpoints of the string list.

While the pointers don't meet, if the lowercase character at the left pointer and the lowercase character at the right pointer are both in the vowel set, swap them and increment/decrement the left and right pointers, respectively. If the lowercase of the left pointer character is not in the vowel set, increment the left pointer, and similarly, if the lowercase of the right pointer is not in the vowel set, decrement the right pointer.

Finally, return the joined string.