Create solution.py
O(n) time and O(n) space solution using a list of the characters in the string and two pointers.

First, convert the string to a list of characters. Initialize the left and right pointers to the endpoints.

Set the offsets of the left side as the ord() of the lowercase 'a' and the ord() of the lowercase 'z' (which is just ord('a') + 25).

While the pointers don't meet, if the lowercase of the left pointer character and the right pointer character are both in this range (inclusive), swap the characters and increment/decrement the left/right pointers. If either the left or right pointer lowercase characters are outside this range, increment/decrement the left/right pointers, respectively.

Return the joined string.