Create solution.py
O(n) time and O(n) space solution using a character list of the string and two pointers.

First, convert the string to a list of characters.

Set the left and right pointers to 0, and a prefix index to -1. The prefix index while get the last index of the prefix (if there is one).

While the right pointer does not go out of bounds of the length of the string, if the word at the right pointer equals the character, set the prefix index to this right pointer and break out of the loop. Otherwise, increment the right pointer.

While the prefix index is not -1 (initial condition) and the left pointer does not meet the prefix index, swap the left pointer and prefix index pointer elements. Then, increment the left pointer and decrement the prefix index.

Return the joined string.