Create solution_bit.py
O(n) time and O(1) space bit manipulation solution.

First get the length of the array.

Initialize the result array, and the x, y values which will be the duplicate and missing values (these are not known ahead of time which one it will be). Also, initialize the variable for XORing.

Go through the range of all numbers from 1 to n, and also the elements in the array. XOR these numbers with itself (which starts with the value 0). The variable will eventually contain the value of xor = x ^ y, where x and y are one of the duplicate and missing.

Now, get the differing bit by starting with 1, and keep seeing if the logical AND of the XORed value x ^ y and the differing bit is not 1, bitshift left.

Then, for the number in the array as well as the numbers in the range 1 through n, if the AND with the differing bit is not 0, then XOR that number with x, otherwise XOR that number with y.

Then, in order to identify the duplicate, run through the array again. If the number equals x (this means x is the duplicate), assign the result appropriately and return. If the number equals y, then y is the duplicate so modify the result array and return.