Create solution_heap.py
O(n + k log n) solution using heap, where k may be n, so O(n + n log n) = O(n log n). This can also be achieved with sorting, but generally, k may be much less than n. Storage is O(n) for the map and the heap.

First, get a frequency map of the numbers in the array.

Create an array of the values of the frequency map, then heapify it.

Let the result initially be the length of the frequency map.

While k is not 0 and the frequency heap is not empty, get the frequency from the min heap by popping. If the k is greater than or equal to the frequency, that means that element may be deleted. Decrement k by this frequency, and decrease the result by 1 (one element has been removed). This is a greedy approach using the minimum frequency elements that can be removed, given the k value.

Finally, return the result.