Create solution.py
O(n) time and O(n) space solution using an array and a map.

First, get a frequency map by counting the number of occurrences of each value in the array.

Create the frequency list which maps the index as the frequency, and how many occurrences of that particular frequency occurs.

For the number-frequency pair in the frequency map, use the frequency as the index and increment it for each occurrence.

Set the result initially to the length of the frequency map.

For the frequency from 1 to the end of the frequency list, set the amount to remove as the element at the particular index equal to the frequency from the frequency list. If k is greater than or equal to the frequency index X removal amount, decrement k by this value and decrement the result by the removal amount. Otherwise, perform integer division and round down (this is the max amount that can be removed for this particular frequency), then decrement the result by this amount and break from the loop.

Return the result.