Create solution.py
O(n 4^n) time solution with O(n) recursion depth for backtracking.

First, initialize the result array.

Create the digit to character map of the phone number digits that map to each 3 (or 4) letter combination.

Define the backtrack function taking in index i and the current string so far. If the length of the current string equals that of the digits, append the current string and return. Otherwise, for each character in the string value of the digit-to-character map, backtrack with the incremented index and the current string appended with the character.

If the digits are not empty, backtrack with index 0 and the empty string.

Return the result.