Create solution.py
O(n^n) worst-case time complexity, but with pruning (difference needs to be 1 between each consecutive splits) so this will be much smaller in practice. O(1) space.

First, define the DFS function with the current index and the previous value passed in. If the index is equal to the length of the string, return true as the base case. For j in the range from the index to the end of the string, get the value as the integer version of the string from the index up to j (inclusive). If the value + 1 equals the previous value (the previous value is 1 more than the current value) and the DFS of the next index j+1 with the value passed in is true, return true. If the loop exits, return false.

Now, perform a iterative loop from index 0 up to the 2nd to last element of the string. Get the value from the start to i (inclusive) and convert it to an integer. If the DFS on the next index (i+1) with the value passed in is true, return true. Otherwise, return false outside the loop.