Create solution.py
O(n^2) time solution and O(n) space for going through a list and dividing the number by the possible choices for each digit. This requires going through the list in an double loop, thus the n^2 complexity.

Initialize the result to an empty string, and create a list of numbers from 1 to n.

For i in range n to 1, get the digit that should be next in the string. This is done by dividing k - 1 value (due to the 0-index of the array) and dividing by the factorial (i-1)!. Then, adjust the k by subtracting by digit times the factorial value for the next value. Append the string of the value at the digit position in the array, then remove that value from the list.

Return the result.