Create solution.lead_lag.sql
Using LEAD() and LAG() window functions. First, get the lag over the numbers and lead over the numbers, as well as the lead and lag over the IDs. The condition is where the next ID is greater than the ID by 1, whereas the previous ID is less than the ID by 1, and the previous and subsequent numbers are equal to the current number.