Create solution.sql
Get the department name and employee name and salary, and name them appropriately. The condition to choose it is in the subquery where the employee's department ID and salary are in the max salary grouped by each department, and the employee ID is equal to the department ID. This will give the max (top) salary per department.