Create solution.py
Use a CTE (common table expression) not_banned that contains all users whose users_id are not banned.

Then, select the request_at and assign as Day, and the rounded sum if the status is either cancelled by driver or client, assigned a 1 for theses cases and 0 otherwise, then divided by the total count and rounded to 2 decimal places. These cases are selected when both client ID and driver ID are in the not_banned CTE, and the request is between the dates 2013-10-01 to 2013-10-03. The cases are grouped by the request dates.