Create solution_trie.py
O(n m) time solution where n is the number of words, and m is the max length of a word (m is the depth of the trie). The space complexity is also O(m) due to BFS recursion.

Define a Trie node and Trie classes, with the Trie class modified to perform BFS and store the value of the word at each child level. The result will be updated to the longest value of the string encountered so far.

Create a trie of the words, by inserting them into the trie, then perform BFS to return the longest word that can be formed in a sequential manner.