Create solution_hashset.py
O(n log n + m) time solution using sorting of the n words and m being the total length of all characters in all strings - there is a O(n) space for the hashset.

First, sort the words.

Create a word set and set the result to the empty string.

For each word in the words list, if the word up to the last character is already in the word set, then if its length is greater than that of the result, set the result to this word. Add the word to the word set.

Finally, return the result.