Create solution_countsort.py
O(n + k) time solution, where n is the number of heights in the list input and k is the count buckets for the possible 1 through 100 values.

First, initialize the result and create a count array for each of the heights from 1 to 100 (this has an extra index at 0, to make the mapping more convenient.

For each height in the array, increment the height at the corresponding index in the count array.

Initialize the expected height array.

For the height in range from 1 to 100 inclusive, set the count as the count at that index. Now, loop over and append to the expected array that many times.

For the index in the heights array, if the expected and the heights array values at that index are not equal, increment the result.

Return the result.