Create solution.py
O(n + n log n) time solution using sorting (potentially for a large portion for the ending array) and the counting sort. O(n) space for the count dictionary and the ending array.

First, create the count dictionary as a defaultdict for arr1, and a set of the values in arr2.

Initialize the ending portion array as an empty array.

For each number in array 1, if it is not in the array 2 set, append it to the ending portion. Increment the count dictionary for that number.

Sort the ending portion array afterwards.

Initialize the result array as empty.

For each number in array 2, get this value in the array 1 count dictionary and append the number that amount of times to the result array.

Finally, return the result array extended by the ending portion (which was sorted earlier).