Create solution.py
O(n) time and O(1) space solution. First, notice that all values before the k value only need to be visited a minimum of either the number of tickets at that point or k. If the value comes after the k index, then those need to be visited either the minimum of the value at that index i or value at k minus 1.

Initialize the result, and loop through with the above logic. Add the requisite amount to the result.

Return the result.