Create solution_sort.py
O(n log n) time and O(1) space solution using sorting.

First, sort the array.

Set the index i to 0, and the previous value to -1. Set the total number of elements on the right of the current index i (including itself) to the length of the array.

While the index is less than the length of the array, if the value at this index equals the total number of elements to the right or the total number to the right is between the previous number and the current number, return the total value. Otherwise, while the next index is not out-of-bounds and the number at the current index equals the next index number, increment the index. Finally, update the previous value to the current value, increment the index, and update the total number of elements to the right to be the length of the array minus the current index.

If the loop does not return, return -1 outside the loop.