Create solution.py
O(n) time and space solution using a count array.

First, initialize the count array to have all values from 0 to n = length of array as the index.

For each number in the array, set the index to the minimum of the number itself and the length of the array, and increment the count in that index.

Set the total to the right as 0.

In the range from reversed order starting at n to 0, get the total right as the count added in suffix order. If the index equals the newly updated total to the right, return.

If the loop finishes, return -1.