Create solution_trie.py
Using trie for O(nm) solution for n being the size of the array and m the length of the bit max of 32-bits.

Implement a TrieNode and Trie class, with the Trie having an increase() function to increase the number of bits by shifting from the MSB to LSB and adding it to the children. Similarly, add the findMax() method and check if the 1 - bit is present in the children, then OR it with the shifted value, otherwise go to the children at that bit. Return the result.

In the main function, increase each number in the trie, and then update the result as the max. and return the result.

This solution has an MLE (memory limit exceeded) or sometimes TLE (time limit exceeded) error.