Create solution.py
O(32n) solution for the loop of 32 bits and running O(n) each time.

First, set the result and bitmask to 0.

Starting from the MSB to LSB, OR the mask with the left-shifted index. Get the found set by ANDing the number in the array with the mask and converting to a set. Now, set the start value to the result ORed with the leftshifted index. For each prefix in the found set, if the XOR of the start and prefix is found, then set the result to the start and break.

Finally, return the result.