Create solution.py
O(n^2) time and O(1) space solution.

First, set the result string to empty and result length to 0.

For the indices in the range of the string length, first set the left and right pointers to the same index (for odd-length palindromes) and the left and right pointers to index i and (i+1) for even-length palindromes.

Then, while the left and right pointers are in-bounds and the left and right pointer have the same characters, if the length of the string (from the left pointer to the right pointer) is greater than the current result length, update the result string and the result length. Decrement the left pointer and increment the right pointer.

After the loop is done, return the result.