Create solution.sql
Select the question ID and set as survey_log, from the survey log table and group by 1 and order by the count of the answer ID divided by the count if the action is show then use the question ID otherwise null. Sort this in descending order for the highest answer rate, and ascending for the question ID. Limit only 1 answer.