Create solution.sql
Select the department name and the conditional sum when the department ID is equal and assign as student_number. This is done on the left-joined table of department to student on the ID, where it is grouped by the 1st column (department name) and ordered by the count in descending order, and then by department name in ascending order (alphabetical).