Create solution_cumul_daily.sql
Get the cumulative daily acceptance rate. First, there needs to be two tables for accepted and rejected indicator variables, then the union of those dates is performed. UNION ALL is used between the rejected and accepted tables since this is faster. Similarly, the dates in request and accept dates that are distinct are UNIONed as well. This is left joined to the UNION ALLed subqueries with indicator variables 'r' for rejected, and 'a' for accepted, and then the join is on all dates that are greater than or equal to for the date of table t1, which contains the values of acceptance/rejection of the requests.