Create solution_recur.py
O(n) time and O(h = log n) or worst-case O(n) space due to recursion stack and height of the tree (the tree is full, but not necessarily a balanced binary tree).

First, if either the left or right child is null (since checking either means both are null due to the condition of the tree being a full binary tree), return the value at that node.

If the node has a value of 2 (OR operation), then return logical OR of the function called recursively on the left and right children. Otherwise (this means the node has a value of 3 (AND operation)), return the logical AND of the function called recursively on the left and right children.