Create solution_iter.py
O(n) time and O(h = log n) or O(n) space complexity from the depth of the tree represented with a stack using iterative DFS.

First, get the stack to contain the root node.

Initialize the node to value mapping.

While the stack is not empty, pop from the stack and assign as the node. If the node does not have a left child (which also means it does not have a right child since the input tree is full), set the value at the node key in the dictionary to whether the node value equals 1 (true if it is, false otherwise). Otherwise, if the left (and right) children are already in the dictionary, if the node value is 2, then set the node value to the logical OR of the left and right children, and if the node value is 3, set the node value in the dictionary to the logical AND of the left and right children. Otherwise, if the nodes are not in the dictionary, extend the stack with the node, the left child, and the right child (in that order). This is to ensure postorder traversal (when going in reverse of the stack).

Finally, return the value in the dictionary at the root node key.