Create solution.py
O(n^2) time solution with O(26) = O(1) space for the characters in each substring.

First, initialize the result to 0.

For i in the range of the length of the string, create a frequency counter for each of the 26 lowercase letters initialized to 0. 
For j in the range from i to the end of the string, increment the frequency counter position (subtracting the appropriate ord() value offset relative to 'a'), and update the result as the max of the frequency counter minus the minimum of the frequency counter (if the value is not 0).

Finally, return the result.