Create solution.sql
Create a CTE for the requester ID, and accepter ID for the number of friends. Then, create a 3rd CTE with the UNION ALL of those two CTEs.

Then, select the ID and sum of the number of friends from the 3rd CTE, group by the ID and order by the number in descending order limiting by 1 output.