Create solution.py
O(n) time and O(n) space solution using BFS with a queue.

First, in the serialize() method, if the root is null return an empty list. Create queue with the root node and a result array with the root value. While the queue is not empty, pop from the queue and if the node is null, continue to the next iteration; otherwise, for each child in the node's children, append to the queue. Then, append the length of the children array (as part of the serialization) to the result array and extend the result array with each of the values of the children. When this while loop finishes, return the result array.

For the deserialize() method, if the data is empty then return null. Create a root node with the data of the first character in the string (this is the invariant in the serialization), then put the remaining data into a queue. Also, create a queue with the root node. While the queue is not empty, pop the node from the queue and continue if it is null. For each value in the popped value from the data (which describes the length of the children), create a child node from popping from the data again, and append this to the node's children (as well as to the queue). Afterwards, return the root.