Create solution.py
O(n) time and space solution using DFS and iterators.

For encode(), return null if the node is null. Create a function for DFS. Set the temp node with the current node value, then for the right child of this node use the next value in the iterator of the siblings and run DFS recursively. For the left child, run the DFS recursively with iterating on the children of the node, then return the temp node. Run this DFS function on the root node and an empty list.

For decode(), create an inner function to traverse the nodes, taking in the node and the parent. If the node is null, return null. Create a node with the tree node value and an empty list, then if the node has a left child, append to the node the children of the traverse method recursed on the left subtree of the tree node. For the right child, append to the parent's child (which will create the siblings) using the traverse() method recursion. Set the node's children to all values reversed. Return the node. Finally, return traverse on the data with none.