Create solution.py
O(n * m) time solution with O(m) space for the queue length and using a trie, where n is the number of folders and m is the average length of the folder (or max length of folder) in terms of characters.

First, define a Trie class.

Then, set the root to a trie. For each index in the length of the folder, set the current value to the root, and for each children of the folder, advance the pointer to the character of each children. Set the current index to the index of the folder.

Create a queue with the root, and initialize the result array.

While the queue is not empty, pop from the queue. If the index of this node is not negative, append the folder at the node's index to the result array. For each character in the children's keys, if the character is not a slash or the index is negative, append the value at that key.

Finally, return the result.