Create solution_exists.sql
This is an efficient way of checking the condition using exists in the subquery, which is much faster than IN/NOT IN, JOIN, GROUP BY, and implicit CROSS JOIN from multiple table queries like FROM t1, t2. The SELECT statement just selects a dummy constant value 1 since the existence of a row is checked, and the data returned is not important.