Create solution.py
O(n log n) time solution and O(n) space using a max heap (negated minheap in python).

First, initialize an empty heap.

For each index up to the 2nd to last, get the difference with the next element. If the difference is <= 0, then continue (no need to use bricks or ladders). Decrement the difference from the bricks, and push the negated difference to the heap.
If the # of bricks becomes negative, and if the ladders are used up (== 0), then return the current index. Otherwise, decrement one ladder, and pop from the heap the value, negate it, and increment the bricks number (due to using a ladder instead of the bricks).

Finally, if the loop does not return, return the last index.