Create solution.py
O(log n) time for addNum, and O(1) for findMedian functions using a maxheap and minheap O(n) memory.

First, create a lower heap (max heap of the lower half of numbers) and higher heap (min heap of the upper half of numbers) in the init.

In addNum(), by default, heap push the number to the lower heap by default (negating it since this is python). Next, check if the value invariant is kept (both the lower and higher heaps exist, and the value at the top of the lower maxheap is greater than the value at the top of the higher minheap). If so, pop the top value of the lower maxheap and add to the higher minheap. If the heaps are unbalanced (either 2 more in the top or lower, pop the top of the respective heap and put the value to the other heap.

For findMedian(), return the top value of the heap with the heap which has a greater length (making sure to negate the values in the lower maxheap). Otherwise, return the average of the two values in the lower and higher heaps.