Create solution.py
O(n log m log n) time solution with O(n) space for the heap, with n being the number of values in the nums array and m being the maximum possible integer value (5 * 10^4 in this case) due to dividing by 2 (hence the log m).

First, initialize the minheap and set the max value to 0.

Now, start the first loop to append values to the minheap - this will be an O(n log m) loop due to a repeated division by 2. For each value in the nums array, set the value to a temp variable. While the number n is even, keep dividing by 2. Finally, append a tuple of the current n value after dividing (if it was an even number) and the max between the temp value and n times 2. Update the max value to the max between itself and the current n.

Initialize the result to infinity, and heapify the min heap that was just created.

Now, this loop is O(n log m log n) (the time complexity of the whole algorithm). While the lengths of the heap and the nums array are the same, get the number n and the max possible value of n by popping from the heap. Set the result to the min of the result itself and the difference (deviation) between the max value and n. If n is less than the max possible value of n, then push to the heap the doubled value of n (since this would be a possible missed value) and max possible value of n. Similarly, update the max value to the max of itself and n times 2.

Finally, return the result.