Create solution.py
O(n log n) time solution with O(n) space using a min heap. First, enumerate the indices and append to each of the inner lists of the tasks list. Then, sort this list by the key of the first value, which is the enqueue time.

Initialize the result array and the min heap, as well as the index i to be 0 (first index of the first task) and the time to be the value of the first task's enqueue time.

While the min heap is not empty or the index for the tasks is not out of bounds, while the index is not out of bounds and the current time is greater than or equal to the current task's enqueue time, push to the heap the processing time and index of the task, and increment the index. If the heap is empty, set the time to the next task's enqueue time; otherwise, get the processing time and task index by popping from the priority queue, and increment the time by the processing time, and append the task index to the result array.

Finally, return the result.