Create solution.py
O(n) time and space solution using a dictionary lookup.

First, initialize a defaultdict of the sums of numbers with the same price-to-index differential. 

For each index in the range of the size of the prices array, get the differential between the prices (y-value) and the index (x-value). Then, increment this diff key in the sum dictionary by the price.

Finally, return the maximum of these summed values.