Create solution.py
O(n) time for constructor (creating a sorted list that pretty much is already a min heap) and O(n) space for the unreserved seats, and O(log n) time for the reserve/unreserve methods. There is only a need to keep track of 1 heap (the unreserved seats) instead of 2 heaps (both reserved and unreserved). There is no need to do checks for reserve()/unreserve() since the problem states the operations will be valid. Therefore, there is no need to keep track of the 2 heaps of reserved seats and unreserved seats.

First, initialize the class with unreserved values from 1 to n in an array.

In reserve(), return the popped value from the unreserved heap.

In unreserve(), push the seat that is the input to the method.