Create solution.py
O(m log n) time solution, where m is the size of the tasks array and n is the size of the servers array. The memory is O(m + n) for the available/unavailable heap and the result array. The solution uses two heaps for the available and unavailable servers, with the available servers' weight and index in the heap, whereas the unavailable servers have the time the server is freed of any tasks, weight, and the index.

First, initialize the result array of length of tasks with 0's.

Create the available server array, which will have tuples of the weight and index of each server, respectively, and heapify it.

Similarly, initialize the empty unavailable server heap.

Set the time t to 0.

For j in the range of the tasks, get the current time by getting the max value of the current time t and the index j. If the available heap is empty, then set the time to the min value time freed in the unavailable heap (which will be the first value since it is a heap).
While the unavailable heap is not empty and the current time t is greater than or equal to the time the first unavailable server is freed, pop from the heap and get the values. Then, push to the available heap the weight and index of the just-freed server.
Pop from the available heap and assign the weight and index. Set this index to the result array's j-th index, and push to the unavailable heap the current time incremented by the task value at index j, the weight and index.

Finally, return the result.