Create solution_heap.py
O(k log n) time solution, where k is input of the ordinal (k-th largest number), and n is the size of the nums array. O(n) space for the heap.

First, create a max heap by converting the strings in nums array to integers and negating. Then heapify it.

While k is greater than 1, pop from the heap and decrement k.

Finally, return the negation of the top of the max heap converted to a string.