Create solution.py
O(n^n) time solution (in reality more O(n!) due to cutting off backtracking when the queen already exists in the corresponding column/positive diagonal/negative diagonal). O(n^2) space for the chess board.

First, create sets for the column indices, positive diagonal indices (which have the row + column constant), and the negative diagonal indices (which have the row - column constant).

Initialize the result array.
Create the board with all .'s  of an n x n array.

Define the backtrack function with input r (for row). If r equals n, this means the end condition is met, so create a copy of the board by joining each row into a string, and append the copy of this board to the result array and return.
For each column in the range up to n, check if the row/column combination already exists in any of the sets of the columns/positive diagonals/negative diagonals, and continue to the next iteration if so. Otherwise, add the column/row + column/row - column to the respective sets and set the board position at (r, c) to 'Q'. Run the backtrack with the row incremented. Afterwards, reverse the row/column combination by removing from the sets and turning the board position at (r, c) back to a '.'.

Run the backtrack starting at the first row (row index 0), then return the result (which should be updated).